Restructure
Major restructure of device files. Also coppied a lot of changes from HTC 7x30 common and put device more in line with other CM9 device files. Updated kernel 2.6.32.59 with ICS kgsl. More testing is needed on verify. Right now the plan is to roll back to IssacJ's 2.6.32.9 as he is using the 3g mem vmsplit. Still lots of work needed to be done to kernel but device files at this point are very solid. However still need to add in remaing code for gps, 3g, etc...